Rename board and add mixins
